+++ CHANGE_LOG.md
@@ -26,6 +26,13 @@ and this project/module adheres to [Semantic Versioning](https://semver.org/spec
     + par_set_u32
     + par_set_i32
     + par_set_f32    
+    + par_set_u8_fast
+    + par_set_i8_fast
+    + par_set_u16_fast
+    + par_set_i16_fast
+    + par_set_u32_fast
+    + par_set_i32_fast
+    + par_set_f32_fast       
  - Improved RAM consumption space for parameter values
  - On parameter value change callbacks
  - Added new API for getting parameter configurations