+++ CHANGE_LOG.md
@@ -5,13 +5,54 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project/module adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ---
+## V3.0.0 - 26.01.2026
+
+### Added
+ - Introduce new warning when seting parameter value out of range - *WAR_LIMITED*
+ - Introduce new error for mutex when failing to obtain mutex
+ - Added generic *PAR_GET* and *PAR_SET* macros
+ - Added set & get API for individual data type
+    + par_get_u8
+    + par_get_i8
+    + par_get_u16
+    + par_get_i16
+    + par_get_u32
+    + par_get_i32
+    + par_get_f32
+    + par_set_u8
+    + par_set_i8
+    + par_set_u16
+    + par_set_i16
+    + par_set_u32
+    + par_set_i32
+    + par_set_f32    
+    + par_set_u8_fast
+    + par_set_i8_fast
+    + par_set_u16_fast
+    + par_set_i16_fast
+    + par_set_u32_fast
+    + par_set_i32_fast
+    + par_set_f32_fast       
+ - Improved RAM consumption space for parameter values
+ - On parameter value change callbacks
+ - Added new API for getting parameter configurations
+    + par_get_name
+    + par_get_desc
+    + par_get_unit
+    + par_get_access
+    + par_is_persistant
+ - Improved parameter table validation checker
+
+### Changed
+ - Forced all enums to C99 integer types
+
+---
 ## V2.2.0 - 06.12.2024
 
 ### Added
  - Added new API functions
     + par_has_changed
     + par_get_type
-    + par_get_range
     + par_get_range
     + par_set_n_save